move cut from StringUtils to RenderedStringUtils, move constants to PlayerInfoOverlay, introduce RenderEvent.ElementType.PLAYER_TAB_LIST, make tab_overlay texture smaller
